requirements.txt
python-telegram-bot>=22.0
ccxt==3.1.60
pandas>=2.0.3
numpy>=1.26.0
matplotlib>=3.7.2
pandas-ta>=0.3.14b0
pytz==2023.3
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.6.1